+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.0.3-alpha] - 2023-01-31
 
 ### Added
 
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Functionality to check if a user has already voted `hasAlreadyVoted`
 - Functionality to check if a user is able to vote `isAbleToVote`
 - Export UMD version via `@vocdoni/sdk/umd`
+- Deterministic Wallet generation from arbitrary data
 
 ### Fixed